moving asg under build pipeline exclusively.
who owns the asg stack tied to the build?  If the site stack owns the set,
we have to update the entire site stack to remain consistent.  if we
attempt to update the stack ourselves, we have a race condition and it's not
worth not knowing exactly what will happen, which might be hard to determine.

on the other hand, we want the infrastructure stack to own the resources in the
stack created for the asg, and indeed, the stack itself and it's parameters.  but if we
own it, we can't updated outside of the main stack to deploy.

The solution I chose was to pass of ownership of the stack completely onto the build.  this means
passing the parameters for the final stack through the build job, which is messy without additional
tooling, but should work if I can get the kinks out.  I set an environment variable for each parameter
and then put them into a json file with some rather ugly python.  A custom build image could do this more
elagantly.

This means the ownership is very clear, and it should be relatively easy to create stacks and manage them.

downsides:

messy passing of parameters through contexts is almost absurdly complex without cloudformation json blob.
looping ends up being less of a challenge but would greatly bolster reusability.

stack won't update till the next deploy, but it should get the newest asg
stack when it does because its s3 url is baked into the build job as an environment variable.
In the mean time, it would be trivial to kickoff the deploy to push a change such as an image update .

It's possible the stack won't be deleted when the build is gone, but that's easy enought to find and clean up.